fix: rename EventBackButton component
